+++ src/fiuba/algo3/modelo/AlgoMonFuego.java
@@ -7,7 +7,7 @@ import src.fiuba.algo3.modelo.tipo.Fuego;
 
 public class AlgoMonFuego extends AlgoMon {
 
-	public AlgoMonFuego(String nombre, int vidaMaxima, List<Ataque> ataques) {
+	public AlgoMonFuego(String nombre, double vidaMaxima, List<Ataque> ataques) {
 		super(nombre, vidaMaxima, ataques);
 		this.tipo = new Fuego();
 	}
