+++ src/fiuba/algo3/modelo/AlgoMonPlanta.java
@@ -7,7 +7,7 @@ import src.fiuba.algo3.modelo.tipo.Planta;
 
 public class AlgoMonPlanta extends AlgoMon {
 
-	public AlgoMonPlanta(String nombre, int vidaMaxima, List<Ataque> ataques) {
+	public AlgoMonPlanta(String nombre, double vidaMaxima, List<Ataque> ataques) {
 		super(nombre, vidaMaxima, ataques);
 		this.tipo = new Planta();
 	}
